crm-erp-system/backend/requirements-test.txt
# Testing Dependencies
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==20.1.0
coverage==7.3.2
django-test-plus==2.4.0
responses==0.24.1
freezegun==1.2.2